Refactored sui_rpc.py to sui_clients package. References #48
